abolish the worker thread and prepare for bluetooth scanning
The worker thread was a mistake.
Androids api guide adviced threads for handling bluetooth, but i misunderstood the intended usage.
It is now cleaned up and moved to more apropriate places.

I also set up a broadcast receiver so that a bluetooth scan may yield new devices in the list.